BatchScorer --> BatchPredictor
