+++ doc/source/ray-air/getting-started.rst
@@ -85,6 +85,9 @@ Predictors
 .. autoclass:: ray.ml.predictor.Predictor
     :members:
 
+.. autoclass:: ray.ml.predictor.BatchPredictor
+    :members:
+
 .. automodule:: ray.ml.predictors.integrations.xgboost
     :members:
     :show-inheritance:
