+++ python/ray/ml/examples/lightgbm_example.py
@@ -4,8 +4,8 @@ from typing import Tuple
 import pandas as pd
 
 import ray
+from ray.ml.predictor import BatchPredictor
 from ray.ml.predictors.integrations.lightgbm import LightGBMPredictor
-from ray.ml.scorer import BatchScorer
 from ray.ml.train.integrations.lightgbm import LightGBMTrainer
 from ray.data.dataset import Dataset
 from ray.ml.result import Result
@@ -57,16 +57,18 @@ def train_lightgbm(num_workers: int, use_gpu: bool = False) -> Result:
 
 def predict_lightgbm(result: Result):
     _, _, test_dataset = prepare_data()
-    scorer = BatchScorer(LightGBMPredictor, result.checkpoint)
+    batch_predictor = BatchPredictor.from_checkpoint(
+        result.checkpoint, LightGBMPredictor
+    )
 
     predicted_labels = (
-        scorer.score(test_dataset)
+        batch_predictor.predict(test_dataset)
         .map_batches(lambda df: (df > 0.5).astype(int), batch_format="pandas")
         .to_pandas(limit=float("inf"))
     )
     print(f"PREDICTED LABELS\n{predicted_labels}")
 
-    shap_values = scorer.score(test_dataset, pred_contrib=True).to_pandas(
+    shap_values = batch_predictor.predict(test_dataset, pred_contrib=True).to_pandas(
         limit=float("inf")
     )
     print(f"SHAP VALUES\n{shap_values}")
